+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - OCV/OMCV bonuses are no longer active effects. They only work for a specific instant attack. (#1285)(https://github.com/dmdorman/hero6e-foundryvtt/issues/1285)
 - Fixed rare issue where adding some 6e powers to a 5e actor would prevent actor sheet from opening.
+- Support for STR Minimum OCV penalty. (#384)(https://github.com/dmdorman/hero6e-foundryvtt/issues/384)
 
 ## Version 4.0.4